projects/pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langsmith>=0.1.139",
    "openai>=1.54.1",
    "python-dotenv>=1.0.1",
    "quotientai>=0.1.1",
    "packaging",
    "tiktoken>=0.8.0",
    "google-generativeai>=0.8.3",
    "tavily-python>=0.5.0",
    "exa-py>=1.6.0",
]